backend/pyproject.toml
[tool.poetry]
name = "samantha-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = ">=3.10,<=3.11"
vocode = "^0.1.110"
python-dotenv = "^1.0.0"
elevenlabs = "^0.2.18"
fastapi = "0.95.1"
transformers = "^4.30.2"
accelerate = "^0.20.3"
torch = {version = "^2.1.0.dev20230619+cu121", source = "cu121"}
torchvision = {version = "^0.16.0.dev20230619+cu121", source = "cu121"}
torchaudio = {version = "^2.1.0.dev20230619+cu121", source = "cu121"}
sentencepiece = "^0.1.99"
uvicorn = "0.21.1"
huggingface-hub = "^0.15.1"


[[tool.poetry.source]]
name = "cu121"
url = "https://download.pytorch.org/whl/nightly/cu121"
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
